fix(review): preserve outstanding allocation invariants
