+++ src/autoskillit/core/types/AGENTS.md
@@ -42,6 +42,7 @@ Type re-export hub and all typed building blocks for the autoskillit package (IL
 | `_type_plugin_source.py` | `PluginSource` discriminated union: `DirectInstall | MarketplaceInstall` |
 | `_type_tradition_manifest.py` | `TraditionManifest`, `LensEntry`, `DialingConfig` frozen dataclasses with `from_dict`/`from_yaml_path` loaders |
 | `_type_context_admission.py` | Frozen content-free identities, events, records, effects, states, and coverage definitions for context admission |
+| `_type_context_admission_validation.py` | Internal validation primitives shared by the context-admission protocol values |
 
 ## Architecture Notes
